Make CI use new org wide CI
- Update ci to call new org wide ci and pass in appropriate
  input parameters.
- For public testing do nothing for now. There are some issues
  running testcases on the public runners.
- For private testing call the new script to run the Jenkins job
  for the common repo.

The make jenkins that used to run on our private runners also
had some code to remove RHEL machines from the list of tests to
run. This got horribly outdated so until we come up with a better
plan public testing is turned off.

We now have a script to call jenkins jobs via the Jenkins API.
However, returning success or failure from the job may not
currently work. We should work on this if it doesn't.